Major refactoring for second prototype.
Entities Ride, Seat, Passenger, Location added to implementation

HTML pages added or modified to support these changes. 

The code base is not stable at the moment, views like Browse Ride will have to be updated to reflect the changes to the model. Also the pages that are there, are just basic versions that show the functionality, they need to be developed more this week. The ERD is being updated with the specific property names and types and will be available in the next few days, when the code is stable.