Feedback pages added :  view a user's feedback - feedback.html  create new feedback on a user - feedback_create.html both are handled by feedback.py
Added and changed some elements in the css file to properly display the feedback forms
Added feedback icons
Changed the date_for_table filter to not display the day (want to keep it minimal length)

Added a method get_name_tag on RoadMateUser to return a formatted name and feedback score, with link to feedback page and to profile, since we currently use (and repeat) this on nearly every page. As I edit the other pages I will standardize the views to use this method to render the formatted username.

For the comments added to Ride and to RideRequest, the formatting is changed slightly

Message is now abstract, RideMessage, RideRequestMessage, and FeedbackMessage inherit from it since they share similar properties. FeedbackForm has been added to render the choice of feedback value as a list of radio buttons.